pyproject.toml
[project]
name = "email-forwarding-checker"
dynamic = ["version"]
description = "Checks if a configured email forwarding is working"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt==1.*",
    "pyyaml==6.*",
    "schedule==1.*",
]

[project.optional-dependencies]
dev = [ "types-paho-mqtt" ]

[project.scripts]
cli-name = "email_forwarding_checker:__main__"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
